ingestion/requirements.txt
requests~=2.32.5
pandas~=2.2.3
SQLAlchemy~=2.0.46
psycopg2-binary~=2.9.11
python-dotenv~=1.0.1
nbformat~=5.10.4
plotly~=5.24.1
jupyter~=1.0.0
nbconvert~=7.16.4